refactor: reservation 관련 DTO를 dto 패키지로 분리
